requirements.txt
# LLM & Orchestration
anthropic>=0.18.0
langchain>=0.1.0
langgraph>=0.0.20
langsmith>=0.0.87

# Vector DB & Embeddings
chromadb>=0.4.22
openai>=1.12.0

# Web Scraping & Processing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
crawl4ai>=0.1.0

# PDF Processing
pypdf>=4.0.0
pdfplumber>=0.10.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
